requirements.txt
openai
azure-identity 
azure-core